requirements.txt
django==2.2.16
djangorestframework==3.12.4
psycopg2
djangorestframework_simplejwt==4.8.0
PyJWT==2.1.0
Pillow